[utilities] Move _sos back into sos/__init__.py
Moving the `_sos_ definition wrapper for gettext into utilities proved
to be the wrong direction, so this moves it back into sos/__init__.py in
order to preserve the ability to properly import it elsewhere.
