+++ sos/utilities.py
@@ -20,7 +20,6 @@ import glob
 import tempfile
 import threading
 import time
-import gettext
 
 from contextlib import closing
 from collections import deque
@@ -28,17 +27,6 @@ from collections import deque
 # PYCOMPAT
 import six
 
-gettext_dir = "/usr/share/locale"
-gettext_app = "sos"
-
-gettext.bindtextdomain(gettext_app, gettext_dir)
-
-
-def _default(msg):
-    return gettext.dgettext(gettext_app, msg)
-
-
-_sos = _default
 
 def tail(filename, number_of_bytes):
     """Returns the last number_of_bytes of filename"""
